requirements.txt
iris-pex-embedded-python
sqlalchemy-iris
langchain==1.0.5
langchain-community
langchain-ollama
langchain-text-splitters
langchain-iris
streamlit
streamlit-chat
pypdf
pytest
chromadb
fastembed